+++ src/app/stores/index.tsx
@@ -1,6 +1,8 @@
 import ResourceStore from './resourceStore'
 import UIStore from './uiStore'
+let us = new UIStore()
+let rs = new ResourceStore(us)
 export default {
-    resourceStore: new ResourceStore(),
-    uiStore: new UIStore()
+    resourceStore: rs,
+    uiStore: us
 }